+++ packages/editor/embedded-document/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/embedded-document
 
+## 0.4.5
+- Updated dependencies [e7292ab444](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/e7292ab444):
+  - @atlaskit/editor-common@37.0.0
+  - @atlaskit/editor-core@108.0.0
+  - @atlaskit/renderer@46.0.0
+
 ## 0.4.4
 - Updated dependencies [9c0b4744be](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/9c0b4744be):
   - @atlaskit/button@12.0.3